TASK: Inline `pruneAllWorkspacesAndContentStreamsFromEventStream` into CR Maintainer
